refactor: fix name to management
