+++ app/src/main/java/com/example/partymaker/utils/ui/navigation/NavigationManager.java
@@ -7,7 +7,7 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 import com.example.partymaker.R;
 import com.example.partymaker.ui.features.core.MainActivity;
-import com.example.partymaker.ui.features.groups.managment.CreateGroupActivity;
+import com.example.partymaker.ui.features.groups.management.CreateGroupActivity;
 import com.example.partymaker.ui.features.groups.discovery.PublicGroupsActivity;
 import com.example.partymaker.ui.features.auxiliary.profile.EditProfileActivity;
