+++ app/src/main/res/layout/activity_party_create.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     android:orientation="vertical"
-    tools:context="com.example.partymaker.ui.features.groups.managment.CreateGroupActivity">
+    tools:context="com.example.partymaker.ui.features.groups.management.CreateGroupActivity">
 
     <!-- Original content in RelativeLayout -->
     <RelativeLayout
