+++ app/src/main/res/layout/activity_party_settings.xml
@@ -8,7 +8,7 @@
     android:fitsSystemWindows="true"
     android:orientation="vertical"
     android:weightSum="10"
-    tools:context="com.example.partymaker.ui.features.groups.managment.AdminSettingsActivity">
+    tools:context="com.example.partymaker.ui.features.groups.management.AdminSettingsActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
